last level

+++ W_7_SI_1/rectangle.h
@@ -3,10 +3,9 @@
 
 class Rectangle :
     public Shape
-
 {
-	int a=1;
-	int b=1;
+	double a=1;
+	double b=1;
 	std::string name = "rectangle";
 	
 	double calculateArea();
@@ -16,7 +15,7 @@ class Rectangle :
 public:
 	Rectangle() :Shape("Perimeter=2a+2b", "Area=a*b"){
 	};
-	Rectangle(std::string name, int a, int b);
+	Rectangle(std::string name, double a, double b);
 	void toString();
 
 };
